+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- Fixed a bug where variants were not getting a default shipping category set if a shipping no longer made it's category available to the product type. ([#4018](https://github.com/craftcms/commerce/issues/4018))
 ## 5.3.13 - 2025-05-21
 
 - Fixed a bug where the “Recipient”, “BCC’d Recipient”, and “CC’d Recipient” email settings weren’t working properly if set to environment variables. ([#4004](https://github.com/craftcms/commerce/issues/4004), [#4002](https://github.com/craftcms/commerce/issues/4002))