+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+- `\craft\commerce\services\LineItems::saveLineItem` and `\craft\commerce\services\OrderAdjustments::saveOrderAdjustment` no longer throws an exception when a line item is not found. Instead, it returns `false` and logs an error. ([#3901](https://github.com/craftcms/commerce/pull/3901))
+
 ## 5.3.12 - 2025-05-07
 
 - Fixed a PHP error that could occur when calculating discount adjustments. ([#3997](https://github.com/craftcms/commerce/issues/3997))